// frontend/src/app/models/product.ts
export class Product {
    _id : String;
    title : String;
    description : String;
    price: string;
    schedule: string;
    imagePath: String;
    isSelected: boolean;
    fullname: String;
    address: String;
    phonenumber: String;
    state: String;
    status: String;
    userid: String;
}